+++ lib/semian/activerecord_trilogy_adapter.rb
@@ -9,7 +9,7 @@ module ActiveRecord
     class TrilogyAdapter
       ActiveRecord::ActiveRecordError.include(::Semian::AdapterError)
 
-      class SemianError < StatementInvalid
+      class SemianError < ConnectionNotEstablished
         def initialize(semian_identifier, *args)
           super(*args)
           @semian_identifier = semian_identifier
@@ -28,6 +28,39 @@ module Semian
 
     ResourceBusyError = ::ActiveRecord::ConnectionAdapters::TrilogyAdapter::ResourceBusyError
     CircuitOpenError = ::ActiveRecord::ConnectionAdapters::TrilogyAdapter::CircuitOpenError
+
+    QUERY_ALLOWLIST = %r{\A(?:/\*.*?\*/)?\s*(ROLLBACK|COMMIT|RELEASE\s+SAVEPOINT)}i
+
+    # The common case here is NOT to have transaction management statements, therefore
+    # we are exploiting the fact that Active Record will use COMMIT/ROLLBACK as
+    # the suffix of the command string and
+    # name savepoints by level of nesting as `active_record_1` ... n.
+    #
+    # Since looking at the last characters in a string using `end_with?` is a LOT cheaper than
+    # running a regex, we are returning early if the last characters of
+    # the SQL statements are NOT the last characters of the known transaction
+    # control statements.
+    class << self
+      def query_allowlisted?(sql, *)
+        # COMMIT, ROLLBACK
+        tx_command_statement = sql.end_with?("T") || sql.end_with?("K")
+
+        # RELEASE SAVEPOINT. Nesting past _3 levels won't get bypassed.
+        # Active Record does not send trailing spaces or `;`, so we are in the realm of hand crafted queries here.
+        savepoint_statement = sql.end_with?("_1") || sql.end_with?("_2")
+        unclear = sql.end_with?(" ") || sql.end_with?(";")
+
+        if !tx_command_statement && !savepoint_statement && !unclear
+          false
+        else
+          QUERY_ALLOWLIST.match?(sql)
+        end
+      rescue ArgumentError
+        return false unless sql.valid_encoding?
+
+        raise
+      end
+    end
 
     attr_reader :raw_semian_options, :semian_identifier
 
@@ -48,7 +81,7 @@ module Semian
     end
 
     def raw_execute(sql, *)
-      if query_allowlisted?(sql)
+      if Semian::ActiveRecordTrilogyAdapter.query_allowlisted?(sql)
         super
       else
         acquire_semian_resource(adapter: :trilogy_adapter, scope: :query) do
@@ -67,17 +100,17 @@ module Semian
     end
 
     def with_resource_timeout(temp_timeout)
-      if connection.nil?
+      if @raw_connection.nil?
         prev_read_timeout = @config[:read_timeout] || 0
         @config.merge!(read_timeout: temp_timeout) # Create new client with temp_timeout for read timeout
       else
-        prev_read_timeout = connection.read_timeout
-        connection.read_timeout = temp_timeout
+        prev_read_timeout = @raw_connection.read_timeout
+        @raw_connection.read_timeout = temp_timeout
       end
       yield
     ensure
       @config.merge!(read_timeout: prev_read_timeout)
-      connection&.read_timeout = prev_read_timeout
+      @raw_connection&.read_timeout = prev_read_timeout
     end
 
     private
@@ -88,21 +121,6 @@ module Semian
         ActiveRecord::ConnectionFailed,
         ActiveRecord::ConnectionNotEstablished,
       ]
-    end
-
-    # TODO: share this with Mysql2
-    QUERY_ALLOWLIST = Regexp.union(
-      %r{\A(?:/\*.*?\*/)?\s*ROLLBACK}i,
-      %r{\A(?:/\*.*?\*/)?\s*COMMIT}i,
-      %r{\A(?:/\*.*?\*/)?\s*RELEASE\s+SAVEPOINT}i,
-    )
-
-    def query_allowlisted?(sql, *)
-      QUERY_ALLOWLIST.match?(sql)
-    rescue ArgumentError
-      return false unless sql.valid_encoding?
-
-      raise
     end
 
     def connect(*args)